catfile: Rename RequestRevision to ReadInfo/ReadObject
Currently both the object_info_reader.go and object_reader.go use
RequestRevision to request info/object data. As we plan to unify this
interface behind the `--batch-command` funcationality of
git-cat-file(1), we want to have more specific names for each of them.
This would facilitate easier merging of the two files.