operations: Add newline after commit message with git2go Merge
A future commit will add an implementation to use git-merge-tree(1) to
create a merge. In order to do so, we will need to also use
git-commit-tree(1) to create the commit.

There is a difference however, between how git-commit-tree(1) processes
the -m message that is passed in versus how git2go's Merge function
creates the commit. git2go's Merge function does not add a '\n' when
creating the commit message.

In order to have parity between git2go and using git-merge-tree(1),
which we are about to do, add a '\n' after the commit message when
merging with git2go.