catfile: Handle paths with newline
The function `ParseObjectInfo` is used to parse object information which
is barfed by `git-cat-file(1)`. This function didn't consider the
possibility that paths could contain newline.

Since paths with newline are allowed by Git, we should support the same.
Modify `ParseObjectInfo` to support this. To understand this, we need to
consider the output of `git-cat-file(1)` for a path with newline,
consider the file `foo\nbar.txt`.

Success would be `<Object_ID> <Object_Type> <Object_Size>`, whereas a
failure would look like:

```
foo
bar.txt missing
```

So to ensure we parse such outputs correctly, we need to parse output
either till we hit EOF or till we come across the `missing` keyword.

This means that errors which are currently caught could also be part of
the path, here we trust the output from Git and keep reading.

Note that we still don't support the input of paths with newline in
`ObjectInfoReader` or `ObjectContentReader`. The upcoming `ObjectReader`
will provide this functionality.